BugID:23251508: lwip_itoa: fix implicit conversion warning
commit 7577e587be028f8295963c9216366228ad922381
Author: Simon Goldschmidt <goldsimon@gmx.de>
Date:   Mon Jun 18 12:15:37 2018 +0200

+++ components/network/lwip/core/def.c
@@ -235,6 +235,6 @@ lwip_itoa(char* result, size_t bufsize, int number)
     return;
   }
   /*  move from temporary buffer to output buffer (sign is not moved) */
-    memmove(res, tmp, (result + bufsize) - tmp);
+  memmove(res, tmp, (size_t)((result + bufsize) - tmp));
 }
 #endif
